merge: fix lint and refactor
