+++ internal/namespaces/object/v1/custom_bucket_lifecycle_test.go
@@ -2,7 +2,6 @@ package object_test
 
 import (
 	"fmt"
-	"os"
 	"testing"
 
 	"github.com/scaleway/scaleway-cli/v2/core"
@@ -177,17 +176,6 @@ func Test_BucketLifecycleGet(t *testing.T) {
 			deleteFile(testLifecyclePath),
 		),
 	}))
-}
-
-func createFile(path, content string) core.BeforeFunc {
-	return func(ctx *core.BeforeFuncCtx) error {
-		err := os.WriteFile(path, []byte(content), 0o644)
-		if err != nil {
-			return fmt.Errorf("error writing file %s: %w", path, err)
-		}
-
-		return nil
-	}
 }
 
 func createLifecycleConfiguration(bucketName, configurationPath string) core.BeforeFunc {
